requirements.txt
Flask
starkbank
python-dotenv
APScheduler
pytest
pytest-env
coverage